+++ AAEmu.Game/Core/Packets/C2G/CSSkillControllerStatePacket.cs
@@ -1,6 +1,8 @@
 ﻿using AAEmu.Commons.Network;
 using AAEmu.Game.Core.Network.Game;
+using AAEmu.Game.Core.Packets.Debug;
 using AAEmu.Game.Models.Game.Skills;
+using AAEmu.Game.Models.Game.Skills.SkillControllers;
 
 namespace AAEmu.Game.Core.Packets.C2G;
